More rename "loop" to "ring"
