+++ examples/sumo/sugiyama.py
@@ -9,7 +9,7 @@ from flow.core.params import SumoParams, EnvParams, \
     InitialConfig, NetParams
 from flow.core.params import VehicleParams
 from flow.envs.ring.accel import AccelEnv, ADDITIONAL_ENV_PARAMS
-from flow.scenarios.loop import LoopScenario, ADDITIONAL_NET_PARAMS
+from flow.scenarios.ring import RingScenario, ADDITIONAL_NET_PARAMS
 
 
 def sugiyama_example(render=None):
@@ -46,7 +46,7 @@ def sugiyama_example(render=None):
 
     initial_config = InitialConfig(bunching=20)
 
-    scenario = LoopScenario(
+    scenario = RingScenario(
         name="sugiyama",
         vehicles=vehicles,
         net_params=net_params,
