+++ flow/controllers/routing_controllers.py
@@ -7,7 +7,7 @@ from flow.controllers.base_routing_controller import BaseRouter
 
 
 class ContinuousRouter(BaseRouter):
-    """A router used to continuously re-route of the vehicle in a closed loop.
+    """A router used to continuously re-route of the vehicle in a closed ring.
 
     This class is useful if vehicles are expected to continuously follow the
     same route, and repeat said route once it reaches its end.
