+++ flow/envs/ring/__init__.py
@@ -1,1 +1,1 @@
-"""Empty init file to ensure documentation for the loop env is created."""
+"""Empty init file to ensure documentation for the ring env is created."""
