+++ flow/multiagent_envs/__init__.py
@@ -1,8 +1,8 @@
 """Empty init file to ensure documentation for multi-agent envs is created."""
 
 from flow.multiagent_envs.multiagent_env import MultiEnv
-from flow.multiagent_envs.loop.wave_attenuation import \
+from flow.multiagent_envs.ring.wave_attenuation import \
     MultiWaveAttenuationPOEnv
-from flow.multiagent_envs.loop.loop_accel import MultiAgentAccelEnv
+from flow.multiagent_envs.ring.accel import MultiAgentAccelEnv
 
 __all__ = ['MultiEnv', 'MultiAgentAccelEnv', 'MultiWaveAttenuationPOEnv']
