+++ flow/scenarios/__init__.py
@@ -10,15 +10,15 @@ from flow.scenarios.bottleneck import BottleneckScenario
 from flow.scenarios.figure_eight import Figure8Scenario
 from flow.scenarios.grid import SimpleGridScenario
 from flow.scenarios.highway import HighwayScenario
-from flow.scenarios.loop import LoopScenario
+from flow.scenarios.ring import RingScenario
 from flow.scenarios.merge import MergeScenario
-from flow.scenarios.loop_merge import TwoLoopsOneMergingScenario
-from flow.scenarios.multi_loop import MultiLoopScenario
+from flow.scenarios.ring_merge import TwoRingsOneMergingScenario
+from flow.scenarios.multi_ring import MultiRingScenario
 from flow.scenarios.minicity import MiniCityScenario
 
 __all__ = [
     "Scenario", "BayBridgeScenario", "BayBridgeTollScenario",
     "BottleneckScenario", "Figure8Scenario", "SimpleGridScenario",
     "HighwayScenario", "LoopScenario", "MergeScenario",
-    "TwoLoopsOneMergingScenario", "MultiLoopScenario", "MiniCityScenario"
+    "TwoRingsOneMergingScenario", "MultiRingScenario", "MiniCityScenario"
 ]
