+++ tests/fast_tests/test_environments.py
@@ -8,7 +8,7 @@ from flow.core.params import VehicleParams
 from flow.core.params import NetParams, EnvParams, SumoParams, InFlows
 from flow.controllers import IDMController, RLController
 from flow.scenarios import LoopScenario, MergeScenario, BottleneckScenario
-from flow.scenarios.loop import ADDITIONAL_NET_PARAMS as LOOP_PARAMS
+from flow.scenarios.ring import ADDITIONAL_NET_PARAMS as LOOP_PARAMS
 from flow.scenarios.merge import ADDITIONAL_NET_PARAMS as MERGE_PARAMS
 from flow.envs import LaneChangeAccelEnv, LaneChangeAccelPOEnv, AccelEnv, \
     WaveAttenuationEnv, WaveAttenuationPOEnv, MergePOEnv, \
@@ -684,7 +684,7 @@ class TestTestEnv(unittest.TestCase):
         net_params = NetParams(additional_params=LOOP_PARAMS)
         env_params = EnvParams()
         sim_params = SumoParams()
-        scenario = LoopScenario("test_loop",
+        scenario = LoopScenario("test_ring",
                                 vehicles=vehicles,
                                 net_params=net_params)
         self.env = TestEnv(env_params, sim_params, scenario)
