+++ tests/fast_tests/test_scenario_base_class.py
@@ -8,7 +8,7 @@ from flow.core.params import NetParams
 from flow.core.params import VehicleParams
 from flow.core.params import EnvParams
 from flow.core.params import SumoParams
-from flow.scenarios.loop import LoopScenario, ADDITIONAL_NET_PARAMS
+from flow.scenarios.ring import RingScenario, ADDITIONAL_NET_PARAMS
 from flow.envs import TestEnv
 from flow.scenarios import Scenario
