+++ tests/stress_tests/stress_test_start.py
@@ -4,7 +4,7 @@ from flow.core.params import SumoParams, EnvParams, \
     InitialConfig, NetParams
 from flow.core.params import VehicleParams
 from flow.envs.ring.accel import AccelEnv, ADDITIONAL_ENV_PARAMS
-from flow.scenarios.loop import LoopScenario, ADDITIONAL_NET_PARAMS
+from flow.scenarios.ring import RingScenario, ADDITIONAL_NET_PARAMS
 import ray
